Allow caller to specify the value of included ref for bundle generation
Summary: Currently, when generating packfile item stream along with refs to include, the caller can specify which refs they want to include or exclude. However, the value of those refs (bookmarks) always needs to be whatever is known by the server. For regular Git, this is fine. However, in Mononoke we can have Commit Cloud commits that are trying to move a bookmark/ref by pushing some commits. At this point, the caller with the CC commits would have their local ref advanced by few commits while the server would still have the ref at its last known public location. Thus, generating a bundle that represents the yet-to-be-landed state requires a mechanism to specify `refs_to_include` along with the commits that they should be pointing at, instead of the server value.

Reviewed By: mitrandir77

Differential Revision: D50334747

fbshipit-source-id: 768209767cc4208a278cf57dcbfde983936aefdb