Integration tests to highlight we don't honor bookmark category while importing git refs
Summary: In the current `GitImport` implementation, we properly categorize annotated tags but simple tags are incorrectly categorized as branches instead. In `RemoteGitImport` all tags are categorized as branches. We discussed within the team and decided that the notion of category in addition to the bookmarks being stored with `heads/` and `tags/` prefix leads to double-source-of-truth problem. In the long term, we want to get rid of the `category` column entirely. For now, we will just be adding integration tests to highlight that we don't actually honor `category` as a reminder that no future code should rely on this column having the correct value.

Reviewed By: gustavoavena

Differential Revision: D50169818

fbshipit-source-id: 24936da66ab935f99f82249b72a42c0bed2d5d63