Enforce derivation of Git commit before GitDeltaManifest
Summary: Even though Git commits have been added as a dependency for generating `GitDeltaManifest`, the lack of actual `derive_dependency` call for Git commit leads to the commit not getting derived in the process of deriving the manifest. This is inconvenient since the callers expect to derive just `GitDeltaManifest` and have all the necessary data for packing/bundling get derived. Additionally, a `GitDeltaManifest` doesn't make sense without an associated Git commit. This diff enforces the git commit derivation as a prerequisite for `GitDeltaManifest` derivation

Reviewed By: gustavoavena

Differential Revision: D50385146

fbshipit-source-id: ea68c1d4d736a890a2672471c011a59167216d09