Allow git bundles to be created with user provided ref values
Summary: Currently, when creating Git bundles the user can either choose to include specific refs or include all refs known by the server. In either case, the value of the refs (i.e. the commit that they point to), is decided by the server. In certain cases, we would want the bundle creator to provide their own ref pointing to a different location (e.g. BUNDLE_HEAD in case of draft commits from `jf submit`). This diff enables the user to provide such an option

Reviewed By: mitrandir77

Differential Revision: D50411687

fbshipit-source-id: 5f4ba771ce51765496246c029e8b1653b02c1e8b