Remove usages of category to determine bookmark type
Summary:
For importing Git refs, we introduced the `BookmarkCategory` type which is supposed to represent whether an imported bookmark is a branch, tag or note. However, the implementation does not honor these types during import and also it leads to multiple sources of truth (the name is stringly typed to contain the category). The team decided to just rely on the name and ignore category as the first step and then completely remove the category column going forward.

The packfile item stream generator was also relying on the category to determine the type of the bookmark. This diff changes that behaviour to instead rely on the name.

Reviewed By: gustavoavena

Differential Revision: D50333802

fbshipit-source-id: 1d1c64e28a4a4fa45f3d2aad4eba4f8b480cbd06