Pre-fetch implicit deletes for all exported changesets
Summary:
## Problem

The biggest bottleck of the tool, by far, was the commit rewriting phase, because we had to get implicit deletes, which do a lot of manifest lookups.

This seems to be very IO bound and could benefit from concurrency.

**However**, the commits are rewritten sequentially in a `try_fold`, because one can only be rewritten after all its parents have been.

This meant that we were also getting the implicit deletes sequentially.

## This diff
This will pre-fetch the implicit deletes for all changesets that will be exported. Since there's no dependency on the parent, this can be done with a high degree of concurrency, which significantly speeds things up.

In the process, I extracted the code that creates the progress bar and added a bar for the pre-fetching stage.

With D50413614 on top, I exported `fbcode/eden/mononoke` to a git repo in 1h11min, down from the +8h that took before.

Reviewed By: RajivTS

Differential Revision: D50410534

fbshipit-source-id: c72503ec8d00afa782ec55664db16ab0a6835b93