Upload and derive changesets as they are created
Summary:
With D50410534, I'll pre-fetch the implicit deletes, so the stage where I copy the changesets is much quicker.

I ran the export of a big directory without this change and noticed that the derivation of the git data types was taking a significant amount of time. So I figured the commits could be uploaded and derived as they're being created.

This really sped things up for big directories.

Reviewed By: RajivTS

Differential Revision: D50413614

fbshipit-source-id: b8cdc7c423b47d6f00f3adc036937031e6feee38