Allow bypass of repo lock at the bookmark level
Summary:
The `create`, `move` and `delete` bookmarks operations are guarded by `repo_lock` (among other things), i.e. if the repo is locked these operations are not permitted. There is a way to bypass this check by providing the `BYPASS_READONLY=true` pushvar. However, in addition to this pushvar, the current code also checks the identities of the user and whether they have the right to provide this pushvar.

Tools like `gitimport` execute from user machines and do not have service certificates for `scm_service_identity` for setting admin level permissions. This is why we use `bypass_access_control` `AuthorizationContext` that is supposed to provide admin level internal-only privileges. This diff updates the check for repo-lock, where we now consider not only the identities of the user but also the authorization context. If any of these indicate that the operation has been initiated by Source Control team or within an Admin context, we allow the bypass

Reviewed By: mzr

Differential Revision: D50495948

fbshipit-source-id: 5466d51aaa36191000ec3e1b9ea48e11c3b9e6c4