Integration test for validating Mononoke draft commit -> Git local commit -> Git landed commit -> Mononoke landed commit workflow
Summary:
With all the work done till this point, we can now have a true E2E test that is the closest possible match to the `whatsapp/server` scenario that we are trying to enable in production. The test has the following steps:
- Create git commits and push them to Git remote
- Due to remote-gitimport, these commits get converted and pushed to Mononoke repo
- Clone the Mononoke repo and add few HG commits. These are commit cloud commits.
- Create Git bundle for these commit cloud commits
- Apply the Git bundle to the original Git repo. Now the Git repo has changes that it can push
- Push the changes to Git remote. Again, remote-gitimport tries to do two things. Convert the Git commits into Mononoke commits and upload them and move the bookmark
- Since the commits are already present at Mononoke end, the first step is skipped. It just moves the bookmark

Note that this diff also includes a code change resolver. This was a round-trippability problem that was encountered by this test. The bonsai commit (BC1) created by importing a Git commit (GC1) contains `committer` field. We also generate HG  commit (HG1) for this bonsai commit and maintain this mapping. When we create a stack in HG (HG2, HG3, HG4...) on top of these imported commits, the client tries to submit the stack with parents (HG1, HG2, HG3, HG4...). At this point Mononoke should just create bonsai changesets for HG2 onwards and skip HG1 since it already exists. However, it still creates a new bonsai changeset for HG1 (BC1*). Now, BC1* != BC1 since BC1 was created via importing GC1 which had committer while BC1* was created from HG1 which does not have the committer field at all. This is where it fails.

This diff adds a check to ensure that we skip recreating and reuploading the bonsai changeset for HG changeset if the HG changeset already has a mapping in `bonsai_hg_mapping` table.

Reviewed By: mitrandir77

Differential Revision: D50470517

fbshipit-source-id: eda7daa1ee26de70a2f3a5efb60d85d3eb84a80f