Ensure that derived thrift Git trees are never different than stored Git trees
Summary:
For commits originating in Mononoke, we would generate both the Thrift Git Tree and the raw Git tree and both of them would have the same hash since we are they generated via the same base. However, in case of Git imported commits in Mononoke, the raw Git tree would already be stored and we would only need to compute the thrift Git tree.

This diff adds a check to validate that the computed thrift Git tree ends up having the same hash as the stored raw Git tree. If these two are different, then it means our code has a bug possibly due to some corner case scenario that we did not anticipate. To ensure that we do not store the faulty trees, we error out whenever we encounter such a situation.

Reviewed By: mitrandir77

Differential Revision: D50495308

fbshipit-source-id: 5fc17acf982e7c2b6d60023bab8a6b41a3d20c7a