optimize FOID creation
Summary:
In the previous diff, we applied FilteredObjectIDs to every tree in the repository regardless of whether any of its descendants were filtered. This was an issue because:

1) Every tree had many bytes of extra information appended to their ObjectID (the path of the tree, the path to the filter file, and a commit hash that pointed to the commit that last modified the filter file)
2) If any FilterId changes occurred, every tree ObjectID changed. That caused ObjectComparisons to return Unknown more frequently
3) These frequent ObjectID changes caused inefficiency in checkout and diff

This diff attempts to remedy those situations by introducing an "unfiltered tree" type of FilteredObjectID. Any tree that has 0 filtered descendants will use this new type. The new type has a few advantages:

1) It's much smaller than a normal Tree FOID. It only contains 1 extra byte that indicates its type
2) Since it doesn't contain any filter information, its ObjectID doesn't change when the filter changes.
3) It gives us more detailed information about the tree and allows us to return ObjectComparison::Different or ObjectComparison::Identical more often (which allows us to avoid the expensive recursive check of all the tree's descendants)

This diff also contains some small changes that were suggested in the last diff.

Reviewed By: xavierd

Differential Revision: D50149393

fbshipit-source-id: 1889ca8ab76c3bf5c62133dcef49061f24acd6c2