do not mark MercurialMatcher as Sync
Summary: We originally marked the MercurialMatcher as Send + Sync. However, this is not true. While it's true that the MercurialMatcher can safely be moved between threads, it's not true that threads can share a reference to a MercurialMatcher. Therefore we should thoroughly document why it's unsafe to do so to prevent any from making the same mistake in the future.

Reviewed By: xavierd

Differential Revision: D50439601

fbshipit-source-id: 5382a8abf5d88a51789dc7ca0bef240d0a35dc43