Start rewriting commits as implicit deletes are fetched
Summary:
Pre-fetching was very effective, but I noticed that I was creating a stream to concurrently pre-fetch, collecting the entire stream and then creating another stream to rewrite the commits sequentially in a fold...

I figured I could chain the two streams together. In the process, to avoid running into the subexecutor problem (not sure if that's the right way to say it markbt 😅), I'm introducing some parallelism to the implicit delete pre-fetching by spawning tokio tasks, which might run some tasks on separate threads.

The only downside I see with this implementation is that I'll have to clone the export paths of every changeset, instead of cloning **a reference to each path** from a changeset.

Given how much faster this runs and that memory usage remained low, I think it's a tradeoff worth making...

Differential Revision: D50493232

fbshipit-source-id: 4b78ef5d8f9fbb1ccbab227a68fec8281c650cd6