progress: add progress bar builder object
Summary:
The main reason I want this is to allow overriding the Registry in tests (i.e. so tests aren't coupled with the global Registry::main()). This way you can set the registry on the builder without having to store it in the progress bar.

This will become more useful as I expand the progress bar API in following diffs.

Reviewed By: quark-zju

Differential Revision: D50510846

fbshipit-source-id: ab7166ab4417cee1b8b8c26feb3634b276a56c13