progress: remove orphan bars before rendering
Summary:
I think this may help avoid jumpiness in some cases.

Previously if a bar finished and a new one started within the same 100ms, the new bar would show up in this render, and then the old bar would disapper in the next render. After this change, the old bar will disappear in the same render the new bar appears.

Reviewed By: quark-zju

Differential Revision: D50510839

fbshipit-source-id: 5c0f0adf5a8f7d17bd762b8bb8fe84a1ef07d22a