Use skeleton manifest to get implicit deletes
Summary:
I briefly discussed this offline with markbt and mitrandir77.
Skeleton manifests are smaller than hg manifests, so they should be a bit cheaper/quicker to parse and perform lookups.

The code in the implicit deletes module is pretty generic for manifests already and required very few changes to use skeleton manifests, so I decided to give it a shot.

The difference is not that big, but there is a difference (see test plan).

## Important notes
The main concern I had when I proposed this was **if all commits being rewritten using the `commit_transformation` library already had skeleton manifests derived**.
If that's the case, then this change should be worth making. Otherwise we could be deriving skeleton manifests unnecessarily.

I asked this and it seems this library is only used with public commits, which all have skeleton manifests derived, so we can proceed with this.

Differential Revision: D50126708

fbshipit-source-id: 535c60daa36b944e7efab75c50a52eaccd563082