service: use VirtualInodeLoader for getSHA1 Thrift call
Summary:
When large amount of SHA1 are requested from EdenFS, these are more than likely
under the same prefixed path, and thus loading a VirtualInode for every path
would reload the same Inode and/or Tree for every path. This is both wasteful
and slow.

The applyToVirtualInode function was designed to avoid exactly this by finding
common prefixes and only loading the VirtualInode once.

Reviewed By: genevievehelsel

Differential Revision: D50284540

fbshipit-source-id: f53dbbd5329e2ff4c5f51381702c27ef3503fdf5