progress: start of "structured" renderer
Summary:
This renderer will reflect the structured nature of progress bars.

Start with some logic to draw nested nature of the bars. We use the fancy unicode lines to be somewhat reminiscent of the (smart)log graph.

The test implements a simplified version of the rendering so we can unit test the pretty unicode art.

Reviewed By: quark-zju

Differential Revision: D50510843

fbshipit-source-id: 3648b81110eb6e3e490fc141af4623694966d8cc