Avoid target graph holes from rust_maybe_universal_binary
Summary:
This makes it so the Mercurial `rust_binary` is always
`//eden/scm:rust_hg`. Then either adds an `alias` or `genrule` for
`//eden/scm:hg`. Changing the name of the `rust_binary` based on config
caused `autocargo-verification` to break in D50510852. This diff also unbreaks
it because the current name is `:rust_hg`.

Reviewed By: quark-zju

Differential Revision: D50582129

fbshipit-source-id: b34c17a8af13d7279f665970f0f43b196383b6b5