Show cwd selector if the current repo is invalid
Summary:
In vscode, we take the first workspace folder as the current repo. This may not be true, and you may have another folder that's an actual sapling repo. So we shoud let you change your cwd selection from the error. Previously, the dropdown was unavailable.

A more ideal behaviour would be to auto-select the first valid sapling repo when opening ISL. This will take a little reshuffling I think, so let's at least start with this to unblock things. This auto-selection could even be done in ISL, by having the list of available cwds be validated first.

Reviewed By: quark-zju

Differential Revision: D50580319

fbshipit-source-id: dc90a6a41249d23573ebcd7c518b9e4e6104a687