Persist drawer state via new platform localStorage api
Summary:
Add an API to the platform to store temporary cachable state. This is useful because it can be accessed synchronously and persistently, and without becoming user-visible as an `sl` config.

The use case in question is to persist the drawer resize and expansion state. So if you resize the commit info, then reload the page, it stays how it was. And same for expanded/collapsed state.

Eventually, we should use this system to persist all sorts of UI expansion state. Right now the commit info sidebar is the only one we can really use.

Reviewed By: quark-zju

Differential Revision: D50571311

fbshipit-source-id: 879e8a42a99960a8220d3ac78cf76fc274aeaef6