Support auto-collapsing the sidebar on resize even when persisted
Summary:
After the last diff, we lost the ability to collapse the sidebar on startup when the window is small, since we'd prefer loading the saved uncollapsed persisted state rather than using the default which has the window size check.

Instead, let's use an effect to set the window-size-dependent collapse. In fact, we should check for auto-collapse any time the window is resized. So if you resize the window to a small size, it is able to auto-collapse the sidebar.

There's a chance this will be annoying, like if you're currently doing something in the sidebar (writing a message or whatever). But all your data is saved, so it would just amount to reopening the window. And the resize probably means you were doing something else anyway.

I also adjusted the size threshold so it auto-collapses a littlem more aggressively.

Reviewed By: sggutier

Differential Revision: D50577174

fbshipit-source-id: 84ebd163dd6ddb08d69d878f1f70df25cf0a89dd