rework the name of the scm:hg buck target yet again
Summary:
Some weeks ago, when we made the `hg` binary a universal binary for macOS we changed the way the `scm:hg` target was generated. This caused some target graph holes, which D50582129 tried to fix. Unfortunately that diff had issues with `run_buck_hgpython.sh`, which is an ugly monster we cannot get rid of yet without doing some significant amount of work.

This diff essentially gives up on having `scm:hg` producing a universal binary for macOS and instead provides a `universal__hg` target that `hgbuild` uses for packaging.

Reviewed By: zertosh

Differential Revision: D50585309

fbshipit-source-id: a938a51c9ec2d5e5899428f94214f0f39b7d3049