inodes: remove QueuedImmediateExecutor from checkout
Summary:
This removes the remaining QueuedImmediateExecutor from checkout, most of the
computation will now run on the Thrift executor which isn't an inline executor
and thus doesn't suffer from deadlocks.

Most of this diff are in tests are these need to be taught to use the
ImmediateFuture instead of the folly::Future.

Reviewed By: genevievehelsel

Differential Revision: D50347460

fbshipit-source-id: 79fcfd2494371c1a9422692862356162b64e11a1