inodes: remove background future from VirtualInodeLoader
Summary:
In D50227150, the QueuedImmediateExecutor was removed from the
VirtualInodeLoader to eliminate a potential deadlock, this however force the
caller to pass down an executor. Instead, a slightly different approach can be
taken: let's simply chain the futures and return an
`ImmediateFuture<vector<T>>`  instead of returning a `vector<SemiFuture<T>>`.

The outcome will be similar: the futures will run on the Thrift executor, but
is now done implicitely in this change.

Reviewed By: genevievehelsel

Differential Revision: D50347463

fbshipit-source-id: cf7023e9eee53e14955b852b44c1902b348dab61