Use local storage to persist getting started modal shown state
Summary:
Previously, we used sl config to remember if you've seen the getting started UI. This doesn't work across remote machines, where you'll have different config.

Instead, we can use local storage, which I added recently as a platform API. This doens't work on all platforms, but it should work in vscode where we use this getting started config anyway.

We need to be backwards compatible, or else we'll show the getting started to everyone next time they load it. This means supporting the sl config as well. But we can write the sl config to local storage, so as long as someone loads this code once, they'll be migrated to the new storage. This means we can delete the backcompat code in a couple of weeks.

Also, while we're at it, we can migrate the name to kekbab-case for standardization.

Reviewed By: quark-zju

Differential Revision: D50625787

fbshipit-source-id: d125e1ad6210f93ec9b71194c875137f0ab7a726