add support for the megarepo forwardsyncer
Summary:
add support for the megarepo forward syncer

this is to differentiate blobstore accesses

we still have lots of unspecified accesses

https://fburl.com/scuba/mononoke_blobstore_trace/345c5pl9

So, adding the megarepo forward syncer would be nice.

Forwardsync calls into the pushrebase logic that calls into Derived Data Service, so we will cover that connection as well.

Reviewed By: quark-zju

Differential Revision: D50600609

fbshipit-source-id: 570ee8cb4f08e180f7d2e2ce8e52c86f53cd6e68