set separate batch_size just for configerator
Summary:
The hg_sync is broken multiple times. Right now it's more than 100 days stale.

Usually we update one task that hosts configerator (due to shardmanager's allocation) with CLI option to decrease batch size and this unblocks the repo, until the next time.

It's not super easy to set this up just for confrigereator as we have one tw job for all repos with SM config.

I'm setting this option here just for configerator. I'm not going to bother making it more general as I think it's not worh it:
- hg sync will be gone some time in the future
- it's only a problem with configereator

Differential Revision: D50690381

fbshipit-source-id: 05b4eb90887b77e3ac510661042619147e4ad2fa