Introduce endpoint for creating Git bundle for a given stack of draft commits
Summary: This diff adds the endpoint for generating Git bundle for a given stack of draft commits and storing that bundle in everstore. As its response, this endpoint returns the everstore handle that can be later used by JellyFish and Landcastle to land these changes. This diff just adds the skeleton, a follow up diff will introduce the core logic for generating and storing this bundle.

Reviewed By: mitrandir77

Differential Revision: D50642618

fbshipit-source-id: fb7556a69d571b1eccb17dea852c334f5e16ef84