Implement generation of bundles/packs with only offset deltas
Summary:
By default, the generator algorithm generates a stream of packfile items with ref deltas as the only form of delta. The reason we pick ref delta is that it allows us to refer to base objects that can show up before or after in the packfile or be present directly at the client. This flexibility allows to use the `commit_ancestor_difference` stream as a stream without collecting it in memory.

On the contrary, offset deltas have a strict requirement that all their base objects should appear before the deltas themselves. This means that we would have to collect and then reverse the `commit_ancestor_difference` stream. It is for this reason that we always prefer ref-deltas. However, in the future there might be scenarios in which we are forced to use just offset deltas. This diff makes it possible to do that.

Reviewed By: markbt

Differential Revision: D50648306

fbshipit-source-id: 37b743b7ca1811d0b7ae011d466c0557256b7696