Allow control over symref inclusion in packfile stream generation
Summary: Currently, the packfile item stream generator produces an output which always includes the `HEAD` symref and the commit that it points to. If the `HEAD` symref is absent, the generator would error out. There can be cases where we want to either include all known symrefs (e.g. git protocol) or include none of them (e.g. jellyfish). This diff adds the necessary changes to make that customization possible

Reviewed By: markbt

Differential Revision: D50645507

fbshipit-source-id: 88aa05b59d13ab478462eac315b56ea3a86d8e04