Migrate Operations to use SucceedableRevset
Summary:
Modify other operations to now take in SucceedableRevset or ExactRevset. This makes it clear that callers should know if they want you to be using a succeedable revset or not.

The new pattern going forward is that any command taking a hash should use SucceedableRevset or ExactRevset.

Maybe we'll make a more convenient name for this union, but right now I think it's ok to be explicit.

Reviewed By: muirdm

Differential Revision: D50674588

fbshipit-source-id: 75bc68f583da0f178e226510ad7cb688e08a38e8