use latestSuccessorUnlessExplicitlyObsolete for hide
Summary:
Like previous diffs, but for HideOperation. This operation was explicitly opted into "exact" revsets before that was even a concept, since it's pretty common to hide an explicitly obsolete commit.

Now that we have a more specific way to deal with exact revsets, let's make things consistent and migrate the HideOperation. It's important we retain the "exact" use case for obsolete commits.

Reviewed By: muirdm

Differential Revision: D50675454

fbshipit-source-id: 1e54b37cd8d8bc4ab57f67068f5988620c0c94d5