Migrate Rebase and Goto to take in SucceedableRevset or ExactRevset
Summary: Make callers be more specific about their requirements—do they want to opt in to automatic successors or not? This lets us do more careful operations. I don't do it in this diff, but later in this stack we can choose to rebase onto an exact commit if it's already obsolete when acted on.

Reviewed By: muirdm

Differential Revision: D50674590

fbshipit-source-id: fe293054bc1a3e5e9ae52b2e055475a65d7de699