Support goto, rebase, and graft in download commits menu
Summary:
A couple of requested features related to downloading commits:
- after downloading, go to that commit
- after downloading, rebase the commit up to the same public base where you are now
- enable grafting commits (notably public commits) on top of the current commit

These can be accomodated by a couple of checkboxes and queueing up some extra commands.

There are options for some of these in `sl pull` and the internal `jf get`. But they have slightly different semantics. I actually prefer having us queue up these commands. You can see the progress more explicitly, and if something goes wrong like conflicts, it's not a pull issue so much as a rebase/graft issue, which is a little easier to understand IMO.

There's an important behavioral distinction if the commit that gets pulled is public or not. When it's not public, the ensueing rebase and goto commands "just work", and we queue them up as usual.

If the pulled commit is public, then rebase doesn't "just work", because rebase errors out if the source cannot be mutated. So we want a way of "copying" instead of "moving". This is basically `graft` / `rebase --keep`. The difference between these two is that graft will automatically `goto`, and `rebase --keep` will not. In fact, it's hard for us to even know the new commit after `rebase --keep`, since the original revset will correspond to the landed commit.

By doing one extra commit lookup after the pull, we can determine if it's public or not, which lets us change our implementation to use rebase/rebase --keep/graft.

This goes against how we typically do such operations. Usually, we just queue up all the things we're going to do. But since the rebase/graft depends on this log, which can't be run until pull finishes, we do not queue these operations. This is a worse UX, but since we already don't know what exactly we're pulling, we don't show optimistic state so it's not really all that much worse.

Reviewed By: quark-zju

Differential Revision: D50250747

fbshipit-source-id: b9049a4bdccc9ae251db9c80c7f9572851eeee7c