Add "ExactRevset" variant
Summary: Expand on the idea of "succeedable-revset" to have an "exact" variant, which describes a revset that should not be automatically succeeded when the operation runs. It will be up the caller of the operation to decide if they want to opt in to successor or not. Some cases, like rebase, depend on the context. For example, rebasing a commit which is visually already obsolete in the UI should use exact and otherwise succeedable... unless it's called from the download commits, where you probably want the exact revset you type in.

Reviewed By: muirdm

Differential Revision: D50674591

fbshipit-source-id: bd7b548e9e8e246d81ded0b964722d71e5773b3a