Include requested refs with value even if they don't exist as bookmarks on the server
Summary: We had extended the functionality of the generator to include user specified refs with value to allow the caller to decide which refs they want in the output bundle. However, we might also want to include refs that are not recognized as bookmarks on the server (e.g. JF requesting an arbitrary ref like `BUNDLE_HEAD` which does not exist). This diff makes the necessary changes to allow such cases.

Reviewed By: mitrandir77

Differential Revision: D50732833

fbshipit-source-id: bb3371f5dee4beac13b6c1ab4bb43959f28e5d02