Changing wording of submit stack modal
Summary:
When confirming the stack submit, we previously had the title phrased as a question. This made it seem like it was a bad idea to submit. But it's not discouraged in any way, we're just using the confirm modal to let you select if you want to submit as a draft or provide an update message.

So let's reword the title to just state what's happening instead of phrase it as a confirmation.

Also, let's put the submit command in the title so it's clear what's being run by this.

Reviewed By: quark-zju

Differential Revision: D50716267

fbshipit-source-id: 866d8f1ad749165bdf747d9a323989bca6f093a8