Collapse Generated file content by default in the comparison view
Summary:
query the generated file status in the comparison view, then partition the files by status so we render generated/partial files below normal files.

Then render a banner describing that this file is generated, with a way to expand it to see the contents anway.

This is a second, parallel notion of "collapse" compared to what we already had. I think this is justified, since it's more like "loading" the genereated file than it is simply collapsing it. This means we don't need to block the comparison view waiting for generated status to come back, and we can easily handle generated status coming in any time.

Reviewed By: quark-zju

Differential Revision: D50624584

fbshipit-source-id: 5b17fe3bb6f2de7749a77b3a201004df063fd2eb