Find status of files as generated or not
Summary:
Initial implementation for generated file detection.

The idea is that genereated files will have `genereated` in their header. We can read files we care about and render them differently on the client. Same for `partially-generated`. This is useful so the UI can deemphasize things like lockfiles or generated code that you don't ever manipulate by hand.

We find this information by reading files we care about directly and looking for the tag in their first 512 bytes.

You could imagine doing this generated file check at every callsite on the server that returns a list of files. So for example, the list of UncommittedChanges would have generatedStatus as one of its built-in fields. This requires slowing down all such fetches by how long it takes to fetch these statuses. Granted, that's usually <50ms, and it's cached, so you only pay once. Nevertheless, I'm worried large changes will see a real slowdown from that. It's already kind of painful if you have dozens of changed files.

Instead, we'll allow the client to store its own cache of generated statuses, and request any paths that it doesn't know about. This way, the generated file statuses don't block anything. It should generally kick in 50-100ms later (less if cached), and do whatever rerendering it needs to. In my testing, this was inperceptible.

TODO: not every file puts this `generated` tag, so we should also support regex-based file detection, and make it customizable (mostly for OSS use cases)

Reviewed By: quark-zju

Differential Revision: D50436452

fbshipit-source-id: 7a30e7eb0646c65512412accd30739661cacc36e