Cache and fetch generated file status on the client
Summary:
Add the client-side cache and send the proper messages to get generated file statuses.

Note that we don't store this in recoil. In order to get recoil rerendering when expected, we instead store a generation number in recoil, which allows all callers to rerender. If you use the provided hooks, this happens behind the scenes for you, along with automatic refetching for files that aren't cached yet.

TODO: better UI for generated files: tooltips like "this file is generated", collapsable section, etc
TODO: use this in other places like comparison view, "tree" mode, split.

Reviewed By: quark-zju

Differential Revision: D50436451

fbshipit-source-id: 467984bc2a5ddf350cbdffcfa9e0fd80fe128c1d