Factor PrjFS on-disk state functions out of WindowsFsck
Summary:
Moves populateOnDiskChildrenState out of the Windows FSCK implementation and
into a separate library target for re-use in the PrjFS background scrubber.

In doing so, this renames newly exported interfaces.  We also add a paramter to
optionally use FIND_FIRST_EX_ON_DISK_ENTRIES_ONLY to be usable while the
virtualization provider is running.

Reviewed By: genevievehelsel

Differential Revision: D50670155

fbshipit-source-id: b8d339da15fe7f3a741168357c7481258c261b66