Implement batch derivation for GitDeltaManifest
Summary: The `GitDeltaManifest` derived data type depends on `GitCommit`, `GitTree` and `Unode` but it doesn't depend on its own parent types. This means that we can heavily parallelize derivation for this type using batch derivation. This diff implements the batch derivation method for `GitDeltaManifest` which I plan to use for backfilling this type in `whatsapp/server` and `aosp/manifest` repo.

Differential Revision: D50789713

fbshipit-source-id: 769c0e70e779f707074fccbaaca87773000a1181