goto: support "--continue"
Summary:
Support "--continue" in the native goto command. Basically, we just read the destination commit hash out of the "updatestate" file. We don't yet support the actual checkout resumption to avoid duplicated work.

I fixed a minor bug in the Python "update" where the "updatemergestate" marker file was being deleted unconditionally if you didn't pass "--continue". This didn't make much of a difference for Python since later code in merge.py would bail out if there were conflicts, but Rust depends on the "updatemergestate" marker's presence to tell if we are in an interrupted "goto --merge".

Reviewed By: quark-zju

Differential Revision: D50570656

fbshipit-source-id: 9adc39f98964d5853ea8686b4a3a1b5ea448b9af