Make "icon" style buttons even more distinct
Summary:
Add a very slight border to appearance="icon" buttons. This makes them stand out just a little bit more, and notably adds 2px of additional width/height which also helps to make them stand out.

This is in response to feedback we've gotten over time that buttons are confusing in the new UI. We made tweaks to this recently, but let's go a little bit further.

Reviewed By: quark-zju

Differential Revision: D50761152

fbshipit-source-id: 3a755d7a85904444411e459d93451ea84fe23677