Fix false positive warning bug
Summary:
The tool should print a warning when there's a chance that an export path is being created in a commit by copying or moving a file from another **not exported** path.

I'm adding a simple test case where we can be 100% sure that the export path is not being created in the commit with the copy_from reference (i.e. commit `C`), because it was already created in commit `A`.

The bug was there because I wasn't updating the `export_paths_not_created` hashset properly when the export paths were being created in changesets without `copy_from` references.

Differential Revision: D50834123

fbshipit-source-id: 00962fa6a4b69b43ac4e77417b76776594513bc2