Refactor integration tests to reduce duplication
Summary:
After D50796027 I noticed that the integration tests started logging to scuba because I set the `mononoke_gitexport` scuba table as the default scuba dataset.

I was going to just append the `--scuba-dataset="file:///tmp/scuba_gitexport_logs"` to all the calls, but that would be a lot of unnecessary repetition and is error-prone for changes in the future.

I figured I would do a small refactor to reduce some existing duplication in the integration tests.

## How
I created a `gitexport_library.sh` bash file with common environment variables and commands run on all gitexport integration tests.

Same as `library.sh`, the integration tests can just import it and run the commands.

Differential Revision: D50878818

fbshipit-source-id: 63a80f294cd3ce4d13613169b13749a402f25ebb