goto: don't bail out on unlink errors
Summary:
On Windows you can't delete files that are open without FILE_SHARE_DELETE specified by the owner (e.g. plain "open()" in Python). The Python checkout code tries various tricks to remove, but if the deletion fails it ultimately issues a warning and proceeds with the checkout. The Rust checkout code has similar tricks, but was bailing out on deletion errors.

In this diff I make deletion errors non-fatal. The checkout library propagates the paths that couldn't be deleted so the caller can issue warnings.

Note that this should unblock experimental.nativecheckout=true on Windows.

Reviewed By: sggutier

Differential Revision: D50758938

fbshipit-source-id: 731d56fc2d29cae0f63eb643ae51760ac0024d95