Prevent deltafication of non-text files and ignore merges
Summary:
The current implementation of `GitDeltaManifest` derivation attempts to find deltas for files up to 250MB in size regardless of whether they are textual or non-textual. Deltas are mostly only useful when code or config or text files get modified in iterations from one commit to the other hence this diff filters out non-textual content for deltafication.

The current GitDeltaManifest implementation performs poorly in case of merges, hence if we encounter a merge (relatively rare), we pick the fulll object instead of the delta.

Reviewed By: mitrandir77

Differential Revision: D50835655

fbshipit-source-id: 4dd6cced0cdd8f3b9a9d12f216a6f229136f27d0